requirements.txt
flask
pandas
openai>=1.0.0
python-dotenv